requirements.txt
astroid==2.2.5
autopep8==1.4.4
AWSIoTPythonSDK==1.4.6
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
dnspython==1.16.0
eventlet==0.25.0
Flask==1.1.1
Flask-JSGlue==0.3.1
Flask-SocketIO==4.2.1
Flask-Uploads==0.2.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
monotonic==1.5
MySQL-python==1.2.3rc1
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.25.0
Pillow==6.0.0
PyAutoGUI==0.9.44
pycodestyle==2.5.0
pycparser==2.19
PyGetWindow==0.0.5
pylint==2.3.1
PyMsgBox==1.0.6
pyparsing==2.4.2
PyRect==0.1.4
PyScreeze==0.1.21
python-dateutil==2.8.0
python-dotenv==0.10.3
python-engineio==3.9.3
python-socketio==4.3.1
PyTweening==1.0.3
pytz==2019.2
pywin32==214
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
sklearn==0.0
typed-ast==1.3.5
urllib3==1.25.3
virtualenv==16.7.5
Werkzeug==0.15.5
wrapt==1.11.1
